implemented mention differentiation between learns and posts
