+++ src/main/java/org/arlevin/memeDatabaseBot/processors/PostMentionsProcessor.java
@@ -1,0 +1,14 @@
+package org.arlevin.memeDatabaseBot.processors;
+
+import lombok.extern.slf4j.Slf4j;
+import org.json.JSONObject;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+public class PostMentionsProcessor {
+
+  void process(JSONObject tweet, String description) {
+
+  }
+}
